made validator global and same validation to user-settings
